+++ common/src/main/java/earth/terrarium/adastra/common/utils/ModUtils.java
@@ -2,16 +2,18 @@ package earth.terrarium.adastra.common.utils;
 
 import earth.terrarium.adastra.common.blockentities.base.ContainerMachineBlockEntity;
 import earth.terrarium.adastra.common.blocks.base.MachineBlock;
+import earth.terrarium.adastra.common.config.AdAstraConfig;
 import earth.terrarium.adastra.common.entities.vehicles.Lander;
 import earth.terrarium.adastra.common.entities.vehicles.Rocket;
+import earth.terrarium.adastra.common.menus.PlanetsMenu;
 import earth.terrarium.adastra.common.menus.base.BaseContainerMenu;
+import earth.terrarium.adastra.common.planets.Planet;
 import earth.terrarium.adastra.common.registry.ModEntityTypes;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.particles.ParticleOptions;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
@@ -29,14 +31,40 @@ public final class ModUtils {
         }
     }
 
+    /**
+     * Gets the machine from a menu packet, if the player is within 8 blocks of the machine and has the menu open.
+     *
+     * @param pos    the position of the machine
+     * @param player the player sending the packet
+     * @param level  the level the machine is in
+     * @return the machine, if the player is within 8 blocks of the machine and has the menu open
+     */
     public static Optional<ContainerMachineBlockEntity> getMachineFromMenuPacket(BlockPos pos, Player player, Level level) {
         if (!(player.containerMenu instanceof BaseContainerMenu<?>))
             return Optional.empty(); // ensure the sender has the menu open
-        if (player.distanceToSqr(pos.getCenter()) > Mth.square(8))
-            return Optional.empty(); // ensure the sender is close enough to the block
+        if (player.distanceToSqr(pos.getCenter()) > 64)
+            return Optional.empty(); // ensure the sender within 8 blocks of the machine
         BlockEntity blockEntity = level.getBlockEntity(pos);
         if (!(blockEntity instanceof ContainerMachineBlockEntity container)) return Optional.empty();
         return Optional.of(container);
+    }
+
+    /**
+     * Checks if the player is able to teleport to a planet. This is successful if the player is in creative, spectator,
+     * OP, or is in a rocket that is above the atmosphere leave config value and with a high enough tier to reach the
+     * target planet.
+     *
+     * @param player       the player to check
+     * @param targetPlanet the planet to teleport to
+     * @return true if the player can teleport to the planet, false otherwise
+     */
+    public static boolean canTeleportToPlanet(Player player, Planet targetPlanet) {
+        if (!(player.containerMenu instanceof PlanetsMenu)) return false;
+        if (player.isCreative() || player.isSpectator() || player.hasPermissions(2)) return true;
+
+        if (!(player.getVehicle() instanceof Rocket rocket)) return false;
+        if (rocket.getY() < AdAstraConfig.atmosphereLeave) return false;
+        return rocket.tier() >= targetPlanet.tier();
     }
 
     public static Direction relative(BlockEntity from, Direction to) {